refactor: move Supabase, Prisma and GitHub domains
