+++ src/domains/supabase/index.ts
@@ -1,5 +1,11 @@
+import { createClient } from "@supabase/supabase-js";
+
 import { randomAlphanumeric } from "@/utils/hooks";
-import { supabaseClient } from "@/utils/supabaseClient";
+
+const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL ?? "";
+const supabaseAnonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY ?? "";
+
+export const supabaseClient = createClient(supabaseUrl, supabaseAnonKey);
 
 export function getSupabaseFileURL(filename: string) {
 	return `${process.env.NEXT_PUBLIC_SUPABASE_STORAGE_URL}/storage/v1/object/public/${filename}`;
